chore: move input store into a dedicated folder
