+++ front/stores/inputStore.tsx
